fix tests and workflows
