+++ .github/workflows/style_docstr.yml
@@ -54,7 +54,7 @@ jobs:
 
       - name: Software Report
         run: |
-          xvfb-run python -c "import pyvista; print(pyvista.Report()); print('Examples path:', pyvista.EXAMPLES_PATH)"
+          xvfb-run python -c "import pyvista; print(pyvista.Report()); from pyvista import examples; print('Examples path:', examples.PATH)"
           which python
           pip list
