+++ .github/workflows/testing-and-deployment.yml
@@ -41,7 +41,7 @@ jobs:
         run: |
           pip install wheel
           pip install .
-          python -c "import pyvista; print(pyvista.Report(gpu=False));print('Examples path:', pyvista.EXAMPLES_PATH)"
+          python -c "import pyvista; print(pyvista.Report(gpu=False));print('Examples path:', pyvista.examples.PATH)"
 
       - name: Install test dependencies
         run: pip install -r requirements_test.txt
@@ -110,7 +110,7 @@ jobs:
           pip install -e .
           which python
           python -c "import pyvista; print(pyvista.Report(gpu=False))"
-          python -c "import pyvista; print(pyvista.EXAMPLES_PATH)"
+          python -c "from pyvista import examples; print(examples.PATH)"
 
       - name: Test Core API
         run: |
@@ -181,7 +181,7 @@ jobs:
 
       - name: Software Report
         run: |
-          xvfb-run python -c "import pyvista; print(pyvista.Report()); print('Examples path:', pyvista.EXAMPLES_PATH)"
+          xvfb-run python -c "import pyvista; print(pyvista.Report()); from pyvista import examples; print('Examples path:', examples.PATH)"
           which python
           pip list
 
@@ -233,12 +233,12 @@ jobs:
         run: |
           pip install wheel
           pip install .
-          python -c "import pyvista; print(pyvista.Report(gpu=False)); print('Examples path:', pyvista.EXAMPLES_PATH)"
+          python -c "import pyvista; print(pyvista.Report(gpu=False)); from pyvista import examples; print('Examples path:', examples.PATH)"
 
       - name: Set example data path as env variable
         run: |
           chcp 65001 #set code page to utf-8
-          echo ("PYVISTA_EXAMPLE_DATA_PATH=" + $(python -c "import pyvista; print(pyvista.EXAMPLES_PATH)")) >> $env:GITHUB_ENV
+          echo ("PYVISTA_EXAMPLE_DATA_PATH=" + $(python -c "import pyvista; from pyvista import examples; print(examples.PATH)")) >> $env:GITHUB_ENV
 
       - name: Cache example data
         uses: actions/cache@v3
