+++ pytest.ini
@@ -14,3 +14,4 @@ testpaths = tests
 markers =
     needs_vtk9: skip test unless VTK is 9.0 or newer.
     needs_vtk_version(version): skip test unless VTK version is at least as specified.
+    skip_download
