+++ pyvista/__init__.py
@@ -44,15 +44,6 @@ warnings.simplefilter(action='ignore', category=FutureWarning)
 # A simple flag to set when generating the documentation
 OFF_SCREEN = os.environ.get("PYVISTA_OFF_SCREEN", "false").lower() == "true"
 
-# If available, a local vtk-data instance will be used for examples
-VTK_DATA_PATH: Optional[str] = None
-if 'PYVISTA_VTK_DATA' in os.environ:
-    VTK_DATA_PATH = os.environ['PYVISTA_VTK_DATA']
-    if not os.path.isdir(VTK_DATA_PATH):
-        warnings.warn(f"VTK_DATA_PATH: {VTK_DATA_PATH} is an invalid path")
-    if not os.path.basename(VTK_DATA_PATH) == 'Data':
-        warnings.warn(f"VTK_DATA_PATH: {VTK_DATA_PATH} is expected to end with 'Data'")
-
 # flag for when building the sphinx_gallery
 BUILDING_GALLERY = False
 if 'PYVISTA_BUILDING_GALLERY' in os.environ:
@@ -64,39 +55,6 @@ REPR_VOLUME_MAX_CELLS = 1e6
 
 # Set where figures are saved
 FIGURE_PATH = None
-
-
-# allow user to override the examples path
-if 'PYVISTA_USERDATA_PATH' in os.environ:
-    USER_DATA_PATH = os.environ['PYVISTA_USERDATA_PATH']
-    if not os.path.isdir(USER_DATA_PATH):
-        raise FileNotFoundError(f'Invalid PYVISTA_USERDATA_PATH at {USER_DATA_PATH}')
-
-else:
-    USER_DATA_PATH = appdirs.user_data_dir('pyvista')
-    try:
-        # Set up data directory
-        os.makedirs(USER_DATA_PATH, exist_ok=True)
-    except Exception as e:
-        warnings.warn(
-            f'Unable to create `PYVISTA_USERDATA_PATH` at "{USER_DATA_PATH}"\n'
-            f'Error: {e}\n\n'
-            'Override the default path by setting the environmental variable '
-            '`PYVISTA_USERDATA_PATH` to a writable path.'
-        )
-        USER_DATA_PATH = ''
-
-EXAMPLES_PATH = os.path.join(USER_DATA_PATH, 'examples')
-try:
-    os.makedirs(EXAMPLES_PATH, exist_ok=True)
-except Exception as e:
-    warnings.warn(
-        f'Unable to create `EXAMPLES_PATH` at "{EXAMPLES_PATH}"\n'
-        f'Error: {e}\n\n'
-        'Override the default path by setting the environmental variable '
-        '`PYVISTA_USERDATA_PATH` to a writable path.'
-    )
-    EXAMPLES_PATH = ''
 
 # Send VTK messages to the logging module:
 send_errors_to_logging()
