+++ pyvista/examples/__init__.py
@@ -1,5 +1,4 @@
-"""Examples routines."""
-
+"""Examples module."""
 from pyvista.examples.downloads import *
 from pyvista.examples.examples import *
 from . import gltf
